+++ OpenWeather-CSharp/Services/DataScienceToolkit/Dto/CoordinatesDto.cs
@@ -2,8 +2,8 @@
 {
     public class CoordinatesDto
     {
-        public float Lat { get; set; }
+        public double Lat { get; set; }
 
-        public float Lon { get; set; }
+        public double Lon { get; set; }
     }
 }
